requirements.txt
asgiref==3.7.2
crispy-bootstrap5==0.7
dj-database-url==2.0.0
Django==4.2.2
django-ckeditor==6.6.1
django-crispy-forms==2.0
django-js-asset==2.1.0
Faker==18.13.0
mysqlclient==2.2.0
psycopg2==2.9.6
psycopg2-binary==2.9.6
python-dateutil==2.8.2
python-decouple==3.8
six==1.16.0
sqlparse==0.4.4
toposort==1.10
typing_extensions==4.6.3
django-cleanup==8.0.0

